Remove global JS
It looks like this JS file does some cosmetic things that we don't rely
on in this repo. Troublingly, it sets some variables on `window` and
breaks if jQuery hasn't been previously loaded.

If we want to re-do this, we should package it as a module and use a
package manager for it rather than loading it from
archive.codeforamerica.org

Also because it is non-TLS-only, it causes our green "https://" to turn
grey.