Remove old /tools folder structure
The files in this folder are comically out-of-date, including a bunch of
stuff for OpenOakland that might have been relevant when starting out,
but is certainly supplanted now by its own website.

Google Analytics shows literally zero traffic to any of these pages,
except for about twenty hits to the recruiting diverse leadership page.
All those hits came from Google, but I'm not sure how because I couldn't
find a link to that page when searching for *exactly* the content of the
page. In my opinion, the technical overhead of maintaining these pages
outweighs their usefulness.