Move almost all JS into one "application" file
This begins a trend to move all JS into an application bundle file,
rather than inserting it dynamically on each page. Child views are still
responsible for calling that page's JS initialization.

I also removed a bunch of JS that seemed unused, and moved the remainder
to the bottom of the <body> (instead of in the <head> and outside the
<body> -- two locations to avoid putting JS!)