Automatically track all internal & external clicks
This adds a global event handler that tracks all clicks. The default
behavior is to track an event with category 'Link Click', action
'click', and label equal to the link's destination href.

However, this can be changed like so:

- if the link is to a different origin, the category changes to
'External Link Click'
- if the link has a "title" attribute, that is used in place of the
"href" attribute for the event's label
- if the link has a "data-analytics-category" attribute, that is used in
place of the default 'Link Click' category (the intent of this is to
allow a product area like the Project Finder to keep its analytics
separate from the rest of the site).

[Closes #469]