Update copy on in-kinds page to match our working doc
After getting a quick round of feedback on the in-kinds page, this
updates the page to have the new draft of content from this doc:

https://docs.google.com/document/d/1cb1d9bRE-ISDbOohyDqr8RZ2APfrmHu2HLZukWrqmGs/edit

There are two sections left commented-out due to uncertainty in our
actual business process for this.

Note that this currently makes the page look pretty bad, but at least
the content is there.