Separate free software into separate pages per product
This converts the mega in-kinds page from a single page into many
smaller pages. Since there isn't an explicit design process we are using
right now, I took a stab at it myself (but happy to have this improved).

The split-up pages will be useful because we can track analytics for
them individually, send links to them easily, and use more complete
descriptions on them.

As part of this, I downloaded a bunch of logos for these products, as
officially as I could find.