WIP Add Webpack
This branch is an experimental take to introduce webpack for compiling
our static assets. Hopefully, if this works, we can get 1) SASS, 2) JS
module stuff, 3) minified images and such, 4) easier integration of
testing tools like jasmine.

This creates a build directory in `brigade/build`. (I was not able to
get the static assets serving properly from the top level of the git
repo.)

TODO:
- [ ] Fix URLs for the PDFs
- [ ] Figure out how to pull in third-party JS without a CDN
- [ ] Write documentation in README
- [ ] Figure out the buildpack necessary for heroku deployment