Add nl2br filter for event location
As of https://github.com/codeforamerica/cfapi/commit/c16ff3e5, the
CfAPI returns location as a string with a newline between the event's
location's name and address (as though representing how you'd write the
event location on an envelope).

By converting the `\n` to a <br> we can display the location properly.
This adds a template method to do just that.