Copy style guide over
An easy way to cut off the dependency on hotlinked files from
style.codeforamerica.org is to just vendor the files into this repo.
Ordinarily (and eventually) we should use a dependency manager for this,
but since the v3 style guide is not going to change any time soon, the
costs of the hotlinking (slower load time for users, more places where
things can go wrong) outweigh the benefits (reusability).

This copies the files over, but webpack is likely to be broken in this
commit. The next commit will make everything work.