Adjust card and card-row so that card-body nolonger uses flexbox
This commit changes cards so that the card body nolonger uses flexbox for its contents, as this was causing some display issues and was probably not the best approach. Flexbox was used previously to display the card action at the end of the card body. Now the card action is seperated out of the card body.

This commit also starts a clean-up of the component stylesheets, alphabetically sorting all properties.