Remove Brigade List
This brigade list does not use any filtering logic to actually display
the correct list of brigades, so it is wrong. It also looks very bad.
And, it pollutes the `/brigade/:id` namespace.

For now, let's convert it to a redirect to the homepage. Some time
later, we can remove it.