Use "bright-v8" instead of default mapbox style layer
The default mapbox styling uses a pure-white (#fff) background for land
areas. User feedback indicated that this makes the navigation controls
hard to see.

By removing the existing tileLayer and adding a "styleLayer" (a class
which extends tileLayer - see [1]), we can make the map a bit more
stylish with a theme that uses a grey for the countries. [Here's a list
of all public styles.][2]

[1]: https://www.mapbox.com/mapbox.js/api/v3.1.1/l-mapbox-stylelayer/
[2]: https://github.com/mapbox/mapbox-gl-styles