Adjust what contact information is displayed when there aren't events
In testing, new Brigade participants said that they would reach out to a
Brigade using their social media accounts, rather than look them up using
the Brigade Directory, as the Brigade Directory felt a little
overwhelming.

This commit adjusts the contact details that are presented to the user
when there are no events, so that the Brigade Directory is only shown as
a last resort.

This commit also correct some typos related to the social media and
contact links in the events section.