Fix robots.txt
The robots.txt file was not being served because the static folder was
not configured in the "brigade" app because of the way that blueprints
get initialized.

By using the copy-webpack-plugin to copy the file and retain a constant
name, we can serve it up from next to the sitemap.