Make get_brigades function return brigades as a list and not a JSON dump
This removes the JSON encoding that was part of the get_brigades function,
moving that encoding to the homepage and map views instead.

The JSON encoding is useful for generating the map, but I'd like to use
the get_brigades function to return a list that I can loop through to
create a brigade list page.