Add Brigade List view showing brigades by state sorted alphabetically
At the moment, the only way to navigate to a brigade profile is through
the map on the home page. During earlier user testing, some participants
mentioned that they would prefer to use a text listing of the brigades
for navigation, or struggled to use the map.

This adds a new page with a text listing of all the official brigades.
Brigades are grouped by State and displayed alphabetically. The State
abbreviation for each brigade is inferred from its City property. When
the City spans two states, as is the case with Code for Kansas City,
it is displayed under both states -- i.e. Kansas and Missouri.