Style cleanup
This commit addresses some style feedback. It reorganizes the style sheet to
move the '@include media' blocks so they're within their parent elements,
helping readability becauese the overrides are closer to the rules of the
parent element. It also puts every rule on a new line, even when a selector
has only a single rule -- this is also in the interest of readability.

See the original feedback on PR #539