Add Community Fellowship pre-header message
To copy the effect from the main CfA webpage -- this creates a big
pre-header link with part of the text underlined to make it obviously
clickable.

This required some additional styles to the v3 style guide's header. I
put them in a new molecule, but it's likely that this molecule will need
changing when we migrate to v4.