Remove /numbers page
Sorry, @themightychris, but we've decided to remove this page for now.

This page keeps coming up in places -- via Google, apparently in old
Knight foundation reports, and in the memories of some longtime brigade
members.

However, the numbers are wrong. They include projects beyond our brigade
network, don't account for the same concepts of "Active brigades" as the
rest of the site, and are all vanity metrics anyway.

Let's reinvent this page to truly capture the scale of the network, and
the local impact the brigades are having in their communities rather
than just pulling some numbers from Github.