Update CARTO instructions
Since we have an enterprise account now, we aren't going to need to
individually apply for CARTO for each brigade that needs it. Instead,
we'll need the email address that the account is registered under.

The project description is still necessary for us at CfA to have
visibility into these projects for storytelling.