Change join_list function to pass through Markup objects
I'd love to use join_list to join a list of HTML links and have the
output marked HTML-safe so the links render. To do this I'd have to call
the function like:

{{ [Markup("<a href=...></a>"), ...] | join_list }}

to designate that the links are safe. Since Markup objects have
overridden common Python string operations (`format` and `join` namely),
to handle the proper escaping, we can get the proper escaping behavior
just by starting with Markup objects in the join_list method.