Update CARTO instructions to reflect product limitations
With apologies to @brandonmichaelgarcia, we have learned over the last
couple months that these instructions were out of date due to the
inability of CARTO to associate existing accounts with the CfA-wide
enterprise account.

As such, the instructions should tell Brigade project leaders to agree
on an email address for the Brigade's CARTO account and then send that
over without creating an existing account.