Fix error caused by empty array in organizations JSON
The website is displaying an error on the homepage: argument of type 'NoneType' is not iterable.
The error involves a variable not being iterable: `org["properties"]["tags"]`. I suspect the error
is caused by an empty 'tags' array (for Code for Ireland) in organizations.json. This commit fixes
the issue by checking that the array is not null before iterating over the contents.