Add dynamic highlighting to nav links
The CfA style guide tends to highlight links by applying an
"active-menu-item" class to the active page. This is a great affordance,
and one we should seek to add to the brigade website redesign MVP.

This adds a new function `nav_link` which handles the logic of
highlighting the link if it is the active page.

I included some janky tests, we'll have to improve this method by having
it highlight the nav item if the user is on a sub-page as well.