Add list-group component
This adds a list-group component, inspired by the Bootstrap component of the same name. It provides
horizontally stacked container for lists of relating things, and each list item can have an action
displayed on the right. I want to use this component to display multiple deploys of the same
project in the Brigade Showcase.