Rename 'alert' component to 'message'
As we're updating the Brigade site to use the v4 styleguide, we realized the  site stylesheets and
tghe v4 styleguide both have an 'alert' component. +tdooner and I discussed it, and
where we've been using the 'alert' component, it's been to provide information rather than to
'alert' the user of anything. Accordingly we agreed to rename the component to 'message'.

This commit also cleans up the SCSS for the message component slightly.