Update styling of message component
Currently the message component looks too similar to the list-group
component that we use on the Project Showcase page, and isn't consistent
with the look-and-feel of the v4 styleg uide. The messages also feels
tooprominent visually. This updates the message component to be a little
more subtle and more consistent with the 'v4' style guide.