Add 'hero_right' block to base template
This adds an area on the right side of the hero component that's a good
place to add a message or a prominent action to the page. For example,
I'd like to use it on the Project Showcase page to add a message asking
users to submit other good projects that they think should be included
on the page.