Add CfA blue and red color variables
This adds the blue and red color variables from the v4 style guide to
the Brigade site style sheets so we can use them. I was getting a
Webpack 'undefined variable' error when using the $color-blue variable
from the v4 style guide in the message component. This addresses that
issue.