Create new stacking context for brigade map
Since the map should be self-contained (i.e. its layers should not be
interspersed between any other layers on the page), we should make the
map its own stacking context. A stacking context is created with a
"position: relative" accompanied by a non-"auto" z-index (see:
https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Positioning/Understanding_z_index/The_stacking_context)
or by a host of other less-applicable methods.

This fixes the problem of the map's navigation appearing atop the mobile
navigation due to both the mobile nav and the map nav existing within
the same context.