Create "subhero" region and update pages to use it
Since putting more and more content into the "hero" was resulting in an
aggressively-red header area, Fritz came up with a strategy of taking
the page's main heading out of the hero and onto the white background of
the page.

I dubbed this section the "subhero" and updated different pages to use
it. Keeping the in-page nav below the hero necessitated the creation of
a new block, in_page_nav, so the in-page-nav components could be
inserted between the hero and subhero.

This commit also contains some visual work for the Brigade pages. I
simplified the logic for the zero-state on Brigade profiles, since we
recently have been strongly encouraging all Brigades to have a Meetup,
it seems less necessary now to support every case of missing information
than it used to.