Disable homepage map zoom until the mouse is moved over map
Ah, the perennial problem with full-width maps: when scrolling past
them, the map captures the scroll event and instead zooms the map.

To prevent this, let's assume that if the user is attempting to use a
scroll wheel to go past the map, they won't be moving the mouse over the
map. This seems to work pretty well in my physical-mouse-based testing.

(I tried the same approach with the 'click' event instead, but that was
a weird interaction -- when I intended to zoom the map, I didn't realize
to click first.)