Add "Other Links" section for Discourse & Fellowship
We wanted to link to these as well, though creating a full card for them
seems a bit odd since they are different than the other resources listed
already. I've broken them into a separate section of other links that
may be useful.

Although this approach doesn't scale for _many_ links, it'll work to get
a few more links on the page and improve the SEO of the Discourse.