Update Donate links to go to newer Donate page
We created this new donate page [seven months ago][1] because the old
donate page required a high maintenance burden and created an
unnecessary duplication in our process.

Now that the old donate page has stopped working for some reason, it's
time to make the switch to the new donate page for real.

[1]: https://groups.google.com/a/codeforamerica.org/g/brigadeleads/c/At0loymFW0I/m/OVC-W90yEAAJ